requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
dashscope==1.23.9
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
MouseInfo==0.1.3
multidict==6.6.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
pytweening==1.2.0
pywin32==311
requests==2.32.4
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
